Refactor ahmia module
