+++ ahmia/ahmia/templates/404.html
@@ -1,8 +1,8 @@
-{% extends "base.html" %}
+{% extends "core.html" %}
 
 {% block title %}Ahmia - 404{% endblock %}
 
-{% block content %}
+{% block body %}
 
 <h2>Not Found</h2>
