+++ ahmia/ahmia/templates/500.html
@@ -1,8 +1,8 @@
-{% extends "base.html" %}
+{% extends "core.html" %}
 
 {% block title %}Ahmia - 500{% endblock %}
 
-{% block content %}
+{% block body %}
 
 <h2>Internal Server Error</h2>
 Error :( Sorry!
