+++ ahmia/ahmia/templates/base.html
@@ -1,4 +1,5 @@
 {% spaceless %}
+{% load staticfiles %}
 
 <!DOCTYPE html>
 <html lang="en">
@@ -9,13 +10,13 @@
   <meta name='keywords' content='onion domains, hidden services, deep web, search engine, onion pages, onion sites, list, cache, tor, tor2web, index, search, pages, show content' />
   <meta name='author' content='ahmia' />
   <title>{% block title %}{% endblock %}</title>
-  <link rel='stylesheet' type='text/css' href='/static/styles/base_styles.css' />
-  <script src='/static/js/jquery.min.js'></script>
-  <script src='/static/js/highlight.js'></script>
-  <link rel="shortcut icon" href="/static/images/favicon.ico" />
-  <link href='/static/styles/fontstyles/Cagliostro.css' rel='stylesheet' />
-  <link href='/static/styles/fontstyles/Bitter.css' rel='stylesheet' />
-  <link href='/static/styles/fontstyles/OpenSans.css' rel='stylesheet' />
+  <link rel='stylesheet' type='text/css' href='{% static "styles/base_styles.css" %}' />
+  <script src='{% static "js/jquery.min.js" %}'></script>
+  <script src='{% static "js/highlight.js" %}'></script>
+  <link rel='shortcut icon' href='{% static "images/favicon.ico" %}' />
+  <link href='{% static "styles/fontstyles/Cagliostro.css" %}' rel='stylesheet' />
+  <link href='{% static "styles/fontstyles/Bitter.css" %}' rel='stylesheet' />
+  <link href='{% static "styles/fontstyles/OpenSans.css" %}' rel='stylesheet' />
 </head>
 
 <body>
@@ -25,9 +26,9 @@
 
    <a href="https://github.com/ahmia/search">
     <img style="position: absolute; top: 0; right: 0; border: 0;"
-      src="/static/images/forkme_right_green_007200.png"
+      src="{% static 'images/forkme_right_green_007200.png' %}"
       alt="Fork me on GitHub"
-      data-canonical-src="/static/images/forkme_right_green_007200.png">
+      data-canonical-src="{% static 'images/forkme_right_green_007200.png' %}">
    </a>
 
     <a href="/" title="ahmia.fi home">
@@ -40,10 +41,17 @@
   </div>
   <hr />
   <footer>
-    <p>This is a search engine for hidden onion sites running inside <a href="https://www.torproject.org/" rel="external">Tor</a> network. Use <a href="https://www.torproject.org/projects/torbrowser.html.en">Tor Browser Bundle</a> to access hidden services.
+    <p>This is a search engine for hidden onion sites running inside
+      <a href="https://www.torproject.org/" rel="external">Tor</a> network. Use
+      <a href="https://www.torproject.org/projects/torbrowser.html.en">
+        Tor Browser Bundle
+      </a> to access hidden services.
       <br />
-      <a href="/add/">Add</a> new sites. <a href="/documentation/indexing/">HS crawling info</a>. <a href="/address/">Inform us</a> about CP and we will <a href="/policy/">filter it</a>.
-      <a href="/stats/viewer">Hidden Websites statistics</a>: <strong class="color12">{{ count_online }} online</strong> &sol; <strong class="color1">{{ count_banned }} filtered</strong>.</p>
+      <a href="/add/">Add</a> new sites. <a href="/documentation/indexing/">HS crawling info</a>.
+      <a href="/address/">Inform us</a> about CP and we will <a href="/policy/">filter it</a>.
+      <a href="/stats/viewer">Hidden Websites statistics</a>:
+      <strong class="color12">{{ count_online }} online</strong> &sol;
+      <strong class="color1">{{ count_banned }} filtered</strong>.</p>
   </footer>
   <nav class="clearfix" role="navigation">
     <div class="item">
@@ -78,7 +86,8 @@
         <ul id="tweet_update_list">
           <li>
             <span>
-	  Juha is enjoying his Google Summer of Code. Developing the search engine in his backyard. <a href="http://t.co/flcMn9s1dj">pic.twitter.com/flcMn9s1dj</a>
+	  Juha is enjoying his Google Summer of Code. Developing the search engine in his backyard.
+    <a href="http://t.co/flcMn9s1dj">pic.twitter.com/flcMn9s1dj</a>
          </span>
           </li>
         </ul>
