+++ ahmia/ahmia/templates/blacklist_report.html
@@ -1,14 +1,16 @@
 {% extends 'core.html' %}
 {% load i18n %}
 {% load ahmia_custom %}
+{% load static %}
+
 {% block title %}{% trans "Hidden Service Blacklist" %} &mdash; Ahmia{% endblock %}
 {% block extra_js %}
   {% is_development_environment as is_dev %}
   {% if is_dev %}
-    <script src="/static/js/blacklist.js"></script>
-    <script src="/static/js/md5.js"></script>
+    <script src="{% static 'js/blacklist.js' %}"></script>
+    <script src="{% static 'js/md5.js' %}"></script>
   {% else %}
-    <script src="/static/blacklist.signed.js"></script>
+    <script src="{% static 'blacklist.signed.js' %}"></script>
   {% endif %}
 {% endblock %}
 {% block body %}
