+++ ahmia/ahmia/templates/createHsDescription.html
@@ -1,9 +1,9 @@
-{% extends "base.html" %}
+{% extends "core.html" %}
 
 {% block title %}Ahmia - Create hs description{% endblock %}
 
-{% block content %}
-
+{% block body %}
+<div id="ahmiaAboutPage">
 <script type="text/javascript">
   function generate() {
     var json = document.getElementById("json");
@@ -129,5 +129,5 @@
   <strong class="color9">http://something.onion/description.json</strong> and our robot will find and use this as an official information about your site. Please make sure that your JSON description is valid JSON. Use exactly the file name description.json
   and be sure that it is accessible under your hidden service /.
 </p>
-
+</div>
 {% endblock %}
