+++ ahmia/ahmia/templates/descriptionProposal.html
@@ -18,7 +18,7 @@
 </p>
 <p>
   <strong>How to do this:</strong>
-  Simply provide a linked description file in your hidden service web page. You can write your information to this <a href="/documentation/createHsDescription/">form</a> and use the generated JSON file. Please make sure that your JSON description is valid
+  Simply provide a linked description file in your hidden service web page. You can write your information to this <a href="{% url 'doc-desc-proposal' %}">form</a> and use the generated JSON file. Please make sure that your JSON description is valid
   JSON. You could then provide JSON information of your hidden service in this way
   <strong class="color9">http://something.onion/description.json</strong>
   and anyone could find this file.
@@ -42,7 +42,7 @@
 </pre>
 
 <p>
-  You can download and study the example <a href="/static/3g2upl4pq6kufc4m.json">JSON file</a>.
+  You can download and study the example <a href="{% static '3g2upl4pq6kufc4m.json' %}">JSON file</a>.
 </p>
 <strong>Help us:</strong>
 <p>
