--- ahmia/ahmia/templates/disclaimer.html
@@ -1,31 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Ahmia - Disclaimer{% endblock %}
-
-{% block content %}
-
-<h2>Disclaimer</h2>
-
-<p>
-  The indexed sites are running inside the Tor network and ahmia.fi is not responsible for them. The administrator of this site (ahmia.fi) cannot be held responsible for what its users post, edit, or any other actions of its users. You may not use this
-  site to edit or add any information when you do not have the legal rights to do so. It is your own responsibility to adhere to these terms and laws when you are using this site.
-</p>
-<p>
-  We cannot guarantee the validity of the information found here. The content of any editable field may recently have been changed and vandalized by anyone.
-</p>
-<p>
-  Our cache is automatically generated by our search crawler software: we do not have anything to do with the actual Tor hidden services content.
-</p>
-<p>
-  This site is fully compliant with the Law of Finland. The administrator is removing any illegal content and is working according to the Finnish law. Please email us and provide exact URL of the content you suspect to be illegal.
-</p>
-
-<h3>Privacy policy</h3>
-<p>
-  We take your privacy seriously: we absolutely do not maintain any IP address logs. We have no information to share to any third parties.
-</p>
-<p>
-  We do not allow backdoors into our services for access by authorities or anyone else. Officials who want information for criminal investigations must enter through our front door with a warrant. If this happens we will publish the warrant and challenge it.
-</p>
-
-{% endblock %}
