+++ ahmia/ahmia/templates/indexing.html
@@ -1,27 +1,30 @@
-{% extends "base.html" %}
+{% extends "core.html" %}
 
 {% block title %}Ahmia - Documentation{% endblock %}
 
-{% block content %}
+{% block body %}
+<div id="ahmiaAboutPage">
+  <h1>Documentation</h1>
+  <h2>Indexing and crawling</h2>
 
-<h2>Indexing and crawling</h2>
+  <p>We are using OnionBot crawler software to build the full text search.</p>
 
-<p>We are using OnionBot crawler software to build the full text search.</p>
+  <p>Furthermore, we are gathering a list of known .onion sites and information about the sites. We try in the following order fill as much as much information about the sites:</p>
 
-<p>Furthermore, we are gathering a list of known .onion sites and information about the sites. We try in the following order fill as much as much information about the sites:</p>
+  <ol>
+    <li>Downloading <a href="{% url 'doc-indexing' %}">the official /description.json</a>
+    </li>
+    <li>Parse information from the HTML of the site</li>
+    <li>Let the users wrote the information</li>
+  </ol>
 
-<ol>
-  <li>Downloading <a href="/documentation/descriptionProposal/">the official /description.json</a>
-  </li>
-  <li>Parse information from the HTML of the site</li>
-  <li>Let the users wrote the information</li>
-</ol>
+  <h3>Robots.txt</h3>
 
-<h3>Robots.txt</h3>
-
-<p>
-  We are honouring robots.txt files. This means that you can prevent our crawler from indexing your page by using robots.txt rules. This way your page is not shown in the search results. On the other hand, your page still shows in the list of all known
-  hidden services that are online.
-</p>
-
+  <p>
+    We are honouring robots.txt files.
+    This means that you can prevent our crawler from indexing your page by using robots.txt rules.
+    This way your page is not shown in the search results.
+    On the other hand, your page still shows in the list of all known hidden services that are online.
+  </p>
+</div>
 {% endblock %}
