+++ ahmia/ahmia/templates/legal.html
@@ -2,8 +2,6 @@
 {% load i18n %}
 {% block title %}{% trans "Legal Disclaimer" %} &mdash; Ahmia{% endblock %}
 {% block body %}
-{% include 'partial/nav_tor.html' with nav_type='maxi' %}
-
 <div id="ahmiaDisclaimer">
   <p>
   <strong>{% trans "Ahmia content warning" %}:</strong>
