--- ahmia/ahmia/templates/login.html
@@ -1,23 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Ahmia - Log in{% endblock %}
-
-{% block content %}
-
-{% if form.errors %}
-<p>Your username and password didn't match. Please try again.</p>
-{% endif %}
-
-<form method="post" action="/rule/login/">
-
-  Username:
-  <input type="text" name="username" size="15" />
-  <br />Password:
-  <input type="password" name="password" size="15" />
-  <br />
-
-  <input type="submit" value="login" />
-  <input type="hidden" name="next" value="{{ next }}" />
-</form>
-
-{% endblock %}
