--- ahmia/ahmia/templates/old_add.html
@@ -1,32 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Ahmia - Hidden service list - Add new domain{% endblock %}
-
-{% block content %}
-
-<h2>Add a new onion domain</h2>
-<p>Please help us to collect more onion domains to this search index.</p>
-<p>URL must be exactly like http://something.onion/</p>
-<form action="/address/" method="post">
-  <fieldset>
-    <p>
-      <label for="add-url">New onion URL:</label>
-      <input type="url" name="url" id="add-url" class="input-text" placeholder="http://something.onion/" required />
-    </p>
-    <input type="submit" value="Add" class="input-add" />
-
-    <noscript>
-      <p>
-        <label for="sign">Type the word antispammer:</label>
-        <br />
-        <input name="sign" id="sign" size="15" />
-      </p>
-    </noscript>
-    <script type="text/javascript">
-      document.write('<input type="hidden" name="sign" value="anti' + 'spammer" />');
-    </script>
-
-  </fieldset>
-</form>
-
-{% endblock %}
