--- ahmia/ahmia/templates/partial/nav_tor.html
@@ -1,29 +1,0 @@
-{% load i18n %}
-<div id="ahmiaHeader" class="{{ nav_type }}">
-  <div id="notTorBrowserWarning" style="display:none">
-    <p>{% trans "Unfortunately we have not deployd non-JavaScript version of Ahmia yet." %}</p>
-  </div>
-  {% if nav_type == 'maxi' %}
-  <div class="maxiContent">
-    <h1><a href="/">Ahmia</a></h1>
-    <form id="searchForm" class="searchForm autocomplete" action="/search/" method="get">
-      <input id="id_q" type="search" name="q" placeholder="{% trans "Search Ahmia" %}" value="{{ query_string }}">
-      <input type="submit" value="{% trans "Search" %}" style="submitButton">
-    </form>
-  </div>
-  {% endif %}
-  <nav>
-    <ul class="primary">
-      <li><a href="/about" title="{% trans "Learn more about Ahmia and its team." %}">{% trans "About Ahmia" %}</a></li>
-      <!-- <li><a href="/directory" title="{% trans "A directory of .onion services." %}">{% trans "Directory" %}</a></li> -->
-      <li><a href="/stats/viewer" title="{% trans "Statistics about the Ahmia engine." %}">{% trans "Statistics" %}</a></li>
-      <li><a href="/add" title="{% trans "Add a hidden service to our index." %}">{% trans "Add Service" %}</a></li>
-      <li><a href="/i2p/" title="{% trans "I2p network search." %}">{% trans "i2p search" %}</a></li>
-    </ul>
-    <ul class="secondary">
-
-      <li><a href="/blacklist" title="{% trans "Blacklist of banned .onion services." %}">{% trans "Blacklist" %}</a></li>
-      <li><a href="/disclaimer">{% trans "Contact" %}</a></li>
-    </ul>
-  </nav>
-</div>
