+++ ahmia/api/views.py
@@ -18,6 +18,12 @@ from ahmia.helpers import (render_page, validate_onion_url, latest_descriptions,
                            redirect_page, get_client_ip)
 
 @require_GET
+def show_ip(request):
+    """Returns the IP address of the request."""
+    ip_addr = get_client_ip(request)
+    return HttpResponse(ip_addr)
+
+@require_GET
 def banned(request):
     """Return the plain text MD5 sums of the banned onions."""
     return banned_txt(request)
@@ -94,43 +100,6 @@ def put_data_to_onion(request, onion):
         print "Error: %s" % onion
         print "Error: %s" % error
         return HttpResponseBadRequest(error)
-
-@require_GET
-def onion_redirect(request):
-    """Add clicked information and redirect to .onion address."""
-    redirect_url = request.GET.get('redirect_url', '')
-    if not redirect_url:
-        answer = "Bad request: no GET parameter URL."
-        return HttpResponseBadRequest(answer)
-    onion = redirect_url.split("://")[1]
-    onion_parts = onion.split(".")
-    for part in onion_parts:
-        if len(part) == 16:
-            onion = part
-            break
-    try:
-        md5 = hashlib.md5(onion+".onion").hexdigest()
-        url = "http://" + onion + ".onion/"
-        validate_onion_url(url)
-        hs, hs_creat = HiddenWebsite.objects.get_or_create(id=onion,
-                                                           url=url,
-                                                           md5=md5)
-        pop, creat = HiddenWebsitePopularity.objects.get_or_create(about=hs)
-        if creat or hs.banned:
-            pop.clicks = 0
-            pop.public_backlinks = 0
-            pop.tor2web = 0
-        pop.clicks = pop.clicks + 1
-        pop.full_clean()
-        pop.save()
-        if hs_creat:
-            hs.full_clean()
-            hs.save()
-    except Exception as error:
-        print "Error with redirect URL: " + redirect_url
-        print error
-    message = "Redirecting to hidden service."
-    return redirect_page(message, 0, redirect_url)
 
 @require_http_methods(['GET', 'PUT'])
 def onion_popularity(request, onion):
